chapter-1/port-scanning/requirements.txt
colorama
python-nmap